refactor(skippy-ffi): split ABI facade by capability
