+++ crates/skippy-ffi/src/lib.rs
@@ -1,1929 +1,132 @@
-pub const ABI_VERSION_MAJOR: u32 = 0;
-pub const ABI_VERSION_MINOR: u32 = 1;
-pub const ABI_VERSION_PATCH: u32 = 39;
-pub const FEATURE_BACKEND_DEVICES: u64 = 1 << 23;
-pub const FEATURE_RUNTIME_EVENTS: u64 = 1 << 24;
-pub const FEATURE_NATIVE_MTP_N1: u64 = 1 << 25;
-pub const FEATURE_NGRAM_CACHE_DRAFT: u64 = 1 << 26;
-pub const FEATURE_INKLING_MTP_MM: u64 = 1 << 27;
-
 #[cfg(feature = "dynamic-runtime")]
 mod dynamic_library;
 
-#[repr(C)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct AbiVersion {
-    pub major: u32,
-    pub minor: u32,
-    pub patch: u32,
-}
-
-/// Whether a native runtime reporting `version` can back this binary's ABI
-/// bindings. Required symbol signatures and by-value struct layouts may change
-/// between patches, so the loader requires an exact ABI match.
-pub const fn runtime_abi_supported(version: AbiVersion) -> bool {
-    version.major == ABI_VERSION_MAJOR
-        && version.minor == ABI_VERSION_MINOR
-        && version.patch == ABI_VERSION_PATCH
-}
-
-use std::ffi::{c_char, c_int, c_void};
-
-pub type LlamaLogCallback =
-    Option<unsafe extern "C" fn(level: c_int, text: *const c_char, user_data: *mut c_void)>;
-pub type MtmdProgressCallback =
-    Option<unsafe extern "C" fn(progress: f32, user_data: *mut c_void) -> bool>;
-pub type SkippyRuntimeEventCallback =
-    Option<unsafe extern "C" fn(event: *const SkippyRuntimeEventV1, user_data: *mut c_void)>;
-
-#[repr(transparent)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct SkippyRuntimeEventCategory(pub u32);
-
-impl SkippyRuntimeEventCategory {
-    pub const MODEL_OPEN: Self = Self(1);
-    pub const BACKEND: Self = Self(2);
-    pub const SESSION: Self = Self(3);
-    pub const KV: Self = Self(4);
-    pub const WARNING: Self = Self(5);
-}
-
-#[repr(transparent)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct SkippyRuntimeEventKind(pub u32);
-
-impl SkippyRuntimeEventKind {
-    pub const MODEL_OPEN_STARTED: Self = Self(1);
-    pub const MODEL_OPEN_PROGRESS: Self = Self(2);
-    pub const BACKEND_DEVICE_SELECTED: Self = Self(3);
-    pub const MODEL_OPEN_FINISHED: Self = Self(4);
-    pub const MODEL_OPEN_FAILED_HANDLED: Self = Self(5);
-}
-
-#[repr(transparent)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct SkippyRuntimeEventEmitterKind(pub u32);
-
-impl SkippyRuntimeEventEmitterKind {
-    pub const UNKNOWN: Self = Self(0);
-    pub const OPEN_THREAD: Self = Self(1);
-    pub const WORKER_THREAD: Self = Self(2);
-}
-
-#[repr(transparent)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct SkippyRuntimeEventProgressUnit(pub u32);
-
-impl SkippyRuntimeEventProgressUnit {
-    pub const NONE: Self = Self(0);
-    pub const BYTES: Self = Self(1);
-    pub const ITEMS: Self = Self(2);
-    pub const TENSORS: Self = Self(3);
-    pub const STEPS: Self = Self(4);
-}
-
-#[repr(transparent)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub struct SkippyRuntimeEventFailureCode(pub u32);
-
-impl SkippyRuntimeEventFailureCode {
-    pub const NONE: Self = Self(0);
-    pub const INVALID_ARGUMENT: Self = Self(1);
-    pub const IO_ERROR: Self = Self(2);
-    pub const MODEL_ERROR: Self = Self(3);
-    pub const RUNTIME_ERROR: Self = Self(4);
-    pub const BACKEND_ERROR: Self = Self(5);
-    pub const CANCELLED: Self = Self(6);
-    pub const INTERNAL_ERROR: Self = Self(7);
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct SkippyRuntimeEventV1 {
-    pub abi_version: u32,
-    pub struct_size: u32,
-    pub category: SkippyRuntimeEventCategory,
-    pub kind: SkippyRuntimeEventKind,
-    pub emitter: SkippyRuntimeEventEmitterKind,
-    pub reserved0: u32,
-    pub sequence: u64,
-    pub timestamp_mono_ns: u64,
-    pub model_id: u64,
-    pub stage_id: u64,
-    pub session_id: u64,
-    pub progress_current: u64,
-    pub progress_total: u64,
-    pub progress_unit: SkippyRuntimeEventProgressUnit,
-    pub failure_code: SkippyRuntimeEventFailureCode,
-    pub status: Status,
-    pub reserved1: u32,
-    pub detail_ptr: *const c_char,
-    pub detail_len: u64,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct SkippyRuntimeEventReporterV1 {
-    pub abi_version: u32,
-    pub struct_size: u32,
-    pub callback: SkippyRuntimeEventCallback,
-    pub user_data: *mut c_void,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum Status {
-    Ok = 0,
-    Error = 1,
-    InvalidArgument = 2,
-    Unsupported = 3,
-    BufferTooSmall = 4,
-    IoError = 5,
-    ModelError = 6,
-    RuntimeError = 7,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum LoadMode {
-    RuntimeSlice = 0,
-    LayerPackage = 1,
-    ArtifactSlice = 2,
-}
-
-#[repr(i32)]
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub enum MtpSource {
-    #[default]
-    Disabled = 0,
-    Integrated = 1,
-    External = 2,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum TensorRole {
-    Unknown = 0,
-    Metadata = 1,
-    Tokenizer = 2,
-    Embedding = 3,
-    Layer = 4,
-    FinalNorm = 5,
-    Output = 6,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum ActivationDType {
-    Unknown = 0,
-    F32 = 1,
-    F16 = 2,
-    Bf16 = 3,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum ActivationLayout {
-    Opaque = 0,
-    TokenMajor = 1,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum BackendDeviceType {
-    Cpu = 0,
-    Gpu = 1,
-    IGpu = 2,
-    Accel = 3,
-    Meta = 4,
-}
-
-pub const BACKEND_DEVICE_CAP_ASYNC: u64 = 1 << 0;
-pub const BACKEND_DEVICE_CAP_HOST_BUFFER: u64 = 1 << 1;
-pub const BACKEND_DEVICE_CAP_BUFFER_FROM_HOST_PTR: u64 = 1 << 2;
-pub const BACKEND_DEVICE_CAP_EVENTS: u64 = 1 << 3;
-
-#[repr(C)]
-pub struct Error {
-    pub status: Status,
-    pub message: *const c_char,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct RuntimeConfig {
-    pub stage_index: i32,
-    pub layer_start: i32,
-    pub layer_end: i32,
-    pub ctx_size: i32,
-    pub lane_count: i32,
-    pub n_batch: i32,
-    pub n_ubatch: i32,
-    pub n_threads: i32,
-    pub n_threads_batch: i32,
-    pub n_gpu_layers: i32,
-    pub has_mmap_override: bool,
-    pub use_mmap: bool,
-    pub use_mlock: bool,
-    pub cache_type_k: i32,
-    pub cache_type_v: i32,
-    pub flash_attn_type: i32,
-    pub load_mode: LoadMode,
-    pub disable_repack: bool,
-    pub use_mmap_prefetch: bool,
-    pub use_mmap_buffer: bool,
-    pub filter_tensors_on_load: bool,
-    pub include_embeddings: bool,
-    pub include_output: bool,
-    pub mtp_source: MtpSource,
-    pub selected_backend_device: *const c_char,
-    pub glm_dsa_policy_profile: i32,
-    pub glm_dsa_policy_flags: u32,
-    pub glm_dsa_short_prefill_max_tokens: i32,
-    pub glm_dsa_direct_sparse_decode_max_top_k: i32,
-    pub glm_dsa_dense_sparse_mask_max_bytes: u64,
-    pub glm_dsa_compact_flash_min_kv: i32,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct BackendDevice {
-    pub version: u32,
-    pub name: *const c_char,
-    pub description: *const c_char,
-    pub device_id: *const c_char,
-    pub memory_free: u64,
-    pub memory_total: u64,
-    pub device_type: BackendDeviceType,
-    pub caps: u64,
-}
-
-#[repr(C)]
-pub struct Model {
-    _private: [u8; 0],
-}
-
-#[repr(C)]
-pub struct NgramCache {
-    _private: [u8; 0],
-}
-
-#[repr(C)]
-pub struct Session {
-    _private: [u8; 0],
-}
-
-#[repr(C)]
-pub struct ModelInfo {
-    _private: [u8; 0],
-}
-
-pub type SkippyModelAttachMtpDraftModelFn = unsafe extern "C" fn(
-    target_model: *mut Model,
-    path: *const c_char,
-    config: *const RuntimeConfig,
-    out_error: *mut *mut Error,
-) -> Status;
-
-pub type SkippyDecodeStepSampledMtpFn = unsafe extern "C" fn(
-    session: *mut Session,
-    token_id: i32,
-    sampling: *const SamplingConfig,
-    out_predicted_token: *mut i32,
-    max_draft_tokens: usize,
-    out_mtp_draft: *mut NativeMtpDraft,
-    out_error: *mut *mut Error,
-) -> Status;
-
-#[repr(C)]
-pub struct SlicePlan {
-    _private: [u8; 0],
-}
-
-#[repr(C)]
-pub struct MtmdContext {
-    _private: [u8; 0],
-}
-
-#[repr(C)]
-pub struct MtmdBitmap {
-    _private: [u8; 0],
-}
-
-#[repr(C)]
-pub struct MtmdInputChunks {
-    _private: [u8; 0],
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum MtmdInputChunkType {
-    Text = 0,
-    Image = 1,
-    Audio = 2,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct MtmdDecoderPos {
-    pub t: u32,
-    pub x: u32,
-    pub y: u32,
-    pub z: u32,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct MtmdInputText {
-    pub text: *const c_char,
-    pub add_special: bool,
-    pub parse_special: bool,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct MtmdContextParams {
-    pub use_gpu: bool,
-    pub print_timings: bool,
-    pub n_threads: c_int,
-    pub image_marker: *const c_char,
-    pub media_marker: *const c_char,
-    pub flash_attn_type: c_int,
-    pub warmup: bool,
-    pub image_min_tokens: c_int,
-    pub image_max_tokens: c_int,
-    pub cb_eval: *mut c_void,
-    pub cb_eval_user_data: *mut c_void,
-    pub batch_max_tokens: c_int,
-    pub progress_callback: MtmdProgressCallback,
-    pub progress_callback_user_data: *mut c_void,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct TensorInfo {
-    pub name: *const c_char,
-    pub layer_index: i32,
-    pub role: TensorRole,
-    pub ggml_type: u32,
-    pub byte_size: u64,
-    pub element_count: u64,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct ActivationDesc {
-    pub version: u32,
-    pub dtype: ActivationDType,
-    pub layout: ActivationLayout,
-    pub producer_stage_index: i32,
-    pub layer_start: i32,
-    pub layer_end: i32,
-    pub token_count: u32,
-    pub sequence_count: u32,
-    pub payload_bytes: u64,
-    pub flags: u64,
-}
-
-pub const ACTIVATION_FLAG_INKLING_MTP_EMBD: u64 = 1 << 2;
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct LogitBias {
-    pub token_id: i32,
-    pub bias: f32,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum LlamaFileType {
-    AllF32 = 0,
-    MostlyF16 = 1,
-    MostlyQ4_0 = 2,
-    MostlyQ4_1 = 3,
-    MostlyQ8_0 = 7,
-    MostlyQ5_0 = 8,
-    MostlyQ5_1 = 9,
-    MostlyQ2K = 10,
-    MostlyQ3KS = 11,
-    MostlyQ3KM = 12,
-    MostlyQ3KL = 13,
-    MostlyQ4KS = 14,
-    MostlyQ4KM = 15,
-    MostlyQ5KS = 16,
-    MostlyQ5KM = 17,
-    MostlyQ6K = 18,
-    MostlyIQ2XXS = 19,
-    MostlyIQ2XS = 20,
-    MostlyQ2KS = 21,
-    MostlyIQ3XS = 22,
-    MostlyIQ3XXS = 23,
-    MostlyIQ1S = 24,
-    MostlyIQ4NL = 25,
-    MostlyIQ3S = 26,
-    MostlyIQ3M = 27,
-    MostlyIQ2S = 28,
-    MostlyIQ2M = 29,
-    MostlyIQ4XS = 30,
-    MostlyIQ1M = 31,
-    MostlyBf16 = 32,
-    MostlyTQ1_0 = 36,
-    MostlyTQ2_0 = 37,
-    MostlyMxfp4Moe = 38,
-    MostlyNvfp4 = 39,
-    MostlyQ1_0 = 40,
-    MostlyQ2_0 = 41,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum GgmlType {
-    F32 = 0,
-    F16 = 1,
-    Q4_0 = 2,
-    Q4_1 = 3,
-    Q5_0 = 6,
-    Q5_1 = 7,
-    Q8_0 = 8,
-    Q8_1 = 9,
-    Q2K = 10,
-    Q3K = 11,
-    Q4K = 12,
-    Q5K = 13,
-    Q6K = 14,
-    Q8K = 15,
-    IQ2XXS = 16,
-    IQ2XS = 17,
-    IQ3XXS = 18,
-    IQ1S = 19,
-    IQ4NL = 20,
-    IQ3S = 21,
-    IQ2S = 22,
-    IQ4XS = 23,
-    I8 = 24,
-    I16 = 25,
-    I32 = 26,
-    I64 = 27,
-    F64 = 28,
-    IQ1M = 29,
-    Bf16 = 30,
-    TQ1_0 = 34,
-    TQ2_0 = 35,
-    Mxfp4 = 39,
-    Nvfp4 = 40,
-    Q1_0 = 41,
-    Q2_0 = 42,
-    Count = 43,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-#[repr(i32)]
-pub enum LlamaModelKvOverrideType {
-    Int = 0,
-    Float = 1,
-    Bool = 2,
-    Str = 3,
-}
-
-#[repr(C)]
-#[derive(Clone, Copy)]
-pub union LlamaModelKvOverrideValue {
-    pub val_i64: i64,
-    pub val_f64: f64,
-    pub val_bool: bool,
-    pub val_str: [c_char; 128],
-}
-
-#[repr(C)]
-#[derive(Clone, Copy)]
-pub struct LlamaModelKvOverride {
-    pub tag: LlamaModelKvOverrideType,
-    pub key: [c_char; 128],
-    pub value: LlamaModelKvOverrideValue,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct LlamaModelTensorOverride {
-    pub pattern: *const c_char,
-    pub tensor_type: GgmlType,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct LlamaModelImatrixData {
-    pub name: *const c_char,
-    pub data: *const f32,
-    pub size: usize,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct LlamaModelQuantizeParams {
-    pub nthread: i32,
-    pub ftype: LlamaFileType,
-    pub output_tensor_type: GgmlType,
-    pub token_embedding_type: GgmlType,
-    pub allow_requantize: bool,
-    pub quantize_output_tensor: bool,
-    pub only_copy: bool,
-    pub pure: bool,
-    pub keep_split: bool,
-    pub dry_run: bool,
-    pub imatrix: *const LlamaModelImatrixData,
-    pub kv_overrides: *const LlamaModelKvOverride,
-    pub tt_overrides: *const LlamaModelTensorOverride,
-    pub prune_layers: *const i32,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy)]
-pub struct SamplingConfig {
-    pub version: u32,
-    pub flags: u32,
-    pub seed: u32,
-    pub top_k: i32,
-    pub penalty_last_n: i32,
-    pub temperature: f32,
-    pub top_p: f32,
-    pub presence_penalty: f32,
-    pub frequency_penalty: f32,
-    pub repeat_penalty: f32,
-    pub logit_bias_count: u32,
-    pub min_p: f32,
-    pub logit_bias: [LogitBias; 256],
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub struct KvPageComponentDesc {
-    pub version: u32,
-    pub role: u32,
-    pub token_start: u64,
-    pub token_count: u64,
-    pub layer_count: u32,
-    pub k_type: u32,
-    pub v_type: u32,
-    pub k_row_bytes: u32,
-    pub v_row_bytes: u32,
-    pub v_element_bytes: u32,
-    pub payload_offset: u64,
-    pub payload_bytes: u64,
-    pub flags: u64,
-}
-
-pub const KV_PAGE_CODEC_SINGLE_V1: u32 = 1;
-pub const KV_PAGE_CODEC_ISWA_COMPOSITE_V1: u32 = 2;
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Default)]
-pub struct KvPageDesc {
-    pub version: u32,
-    pub layer_start: i32,
-    pub layer_end: i32,
-    pub token_start: u64,
-    pub token_count: u64,
-    pub layer_count: u32,
-    pub k_type: u32,
-    pub v_type: u32,
-    pub k_row_bytes: u32,
-    pub v_row_bytes: u32,
-    pub v_element_bytes: u32,
-    pub payload_bytes: u64,
-    pub flags: u64,
-    pub codec: u32,
-    pub component_count: u32,
-    pub components: [KvPageComponentDesc; 2],
-}
-
-pub const NATIVE_MTP_MAX_DRAFT_TOKENS: usize = 8;
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq)]
-pub struct NativeMtpDraft {
-    pub version: u32,
-    pub available: bool,
-    pub token_count: i32,
-    pub token_ids: [i32; NATIVE_MTP_MAX_DRAFT_TOKENS],
-    pub proposal_compute_us: i64,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Default, PartialEq)]
-pub struct TokenSignal {
-    pub entropy: f32,
-    pub top_logprob: f32,
-    pub second_logprob: f32,
-    pub margin: f32,
-    pub top_token: i32,
-    pub second_token: i32,
-}
-
-#[repr(C)]
-#[derive(Debug, Clone, Copy, Default, PartialEq)]
-pub struct GenerationSignalWindow {
-    pub token_count: u32,
-    pub mean_entropy: f32,
-    pub max_entropy: f32,
-    pub mean_margin: f32,
-    pub min_margin: f32,
-    pub high_entropy_count: u32,
-    pub repetition_count: u32,
-}
-
-#[cfg(not(feature = "dynamic-runtime"))]
-/// Mark the statically linked native runtime as already available.
-pub fn native_runtime_loaded() -> bool {
-    true
-}
-
-#[cfg(not(feature = "dynamic-runtime"))]
-/// No-op for statically linked builds.
-///
-/// # Safety
-///
-/// Static builds resolve the native ABI at process link/load time, so this
-/// function does not dereference the supplied path or mutate loader state.
-pub unsafe fn load_native_runtime_library(
-    _path: impl AsRef<std::path::Path>,
-) -> Result<(), NativeRuntimeLoadError> {
-    Ok(())
-}
-
-#[cfg(not(feature = "dynamic-runtime"))]
-/// No-op for statically linked builds.
-///
-/// # Safety
-///
-/// Static builds resolve the native ABI at process link/load time, so this
-/// function does not dereference the supplied paths or mutate loader state.
-pub unsafe fn load_native_runtime_libraries<I, P>(_paths: I) -> Result<(), NativeRuntimeLoadError>
-where
-    I: IntoIterator<Item = P>,
-    P: AsRef<std::path::Path>,
-{
-    Ok(())
-}
-
-#[derive(Debug)]
-pub enum NativeRuntimeLoadError {
-    Load(String),
-    AlreadyLoaded,
-}
-
-impl std::fmt::Display for NativeRuntimeLoadError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Load(message) => write!(f, "{message}"),
-            Self::AlreadyLoaded => write!(f, "native runtime library is already loaded"),
-        }
-    }
-}
-
-impl std::error::Error for NativeRuntimeLoadError {}
-
+mod abi;
+mod activation;
 #[cfg(feature = "dynamic-runtime")]
-mod dynamic {
-    use super::*;
-    use libloading::Library;
-    use std::sync::OnceLock;
-
-    static SYMBOLS: OnceLock<Symbols> = OnceLock::new();
-
-    pub fn native_runtime_loaded() -> bool {
-        SYMBOLS.get().is_some()
-    }
-
-    /// Load a native runtime library and resolve the Skippy ABI symbols from it.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure the library is a MeshLLM native runtime built for
-    /// the running MeshLLM version and Skippy ABI. Loading an arbitrary library
-    /// can bind incompatible C symbols and cause undefined behavior in later FFI
-    /// calls.
-    pub unsafe fn load_native_runtime_library(
-        path: impl AsRef<std::path::Path>,
-    ) -> Result<(), NativeRuntimeLoadError> {
-        let symbols = unsafe { Symbols::load_paths(&[path.as_ref()]) }?;
-        SYMBOLS
-            .set(symbols)
-            .map_err(|_| NativeRuntimeLoadError::AlreadyLoaded)
-    }
-
-    /// Load native runtime libraries and resolve Skippy ABI symbols from them.
-    ///
-    /// Libraries are loaded in the provided order and symbols are resolved by
-    /// searching from the last library backwards, so dependencies should appear
-    /// before the primary `libllama`/`llama.dll` entry.
-    ///
-    /// # Safety
-    ///
-    /// The caller must ensure every library belongs to the same MeshLLM native
-    /// runtime artifact for the running MeshLLM version and Skippy ABI. Loading
-    /// arbitrary libraries can bind incompatible C symbols and cause undefined
-    /// behavior in later FFI calls.
-    pub unsafe fn load_native_runtime_libraries<I, P>(
-        paths: I,
-    ) -> Result<(), NativeRuntimeLoadError>
-    where
-        I: IntoIterator<Item = P>,
-        P: AsRef<std::path::Path>,
-    {
-        let collected = paths
-            .into_iter()
-            .map(|path| path.as_ref().to_path_buf())
-            .collect::<Vec<_>>();
-        let symbols = unsafe { Symbols::load_paths(&collected) }?;
-        SYMBOLS
-            .set(symbols)
-            .map_err(|_| NativeRuntimeLoadError::AlreadyLoaded)
-    }
-
-    fn symbols() -> &'static Symbols {
-        SYMBOLS
-            .get()
-            .expect("MeshLLM native runtime library has not been loaded")
-    }
-
-    type SkippyAbiVersionFn = unsafe extern "C" fn() -> AbiVersion;
-
-    fn check_runtime_abi(libraries: &[Library]) -> Result<(), NativeRuntimeLoadError> {
-        let mut abi_version = None;
-        for library in libraries.iter().rev() {
-            if let Ok(symbol) =
-                unsafe { library.get::<SkippyAbiVersionFn>(b"skippy_abi_version\0") }
-            {
-                abi_version = Some(unsafe { symbol() });
-                break;
-            }
-        }
-        let Some(version) = abi_version else {
-            return Err(NativeRuntimeLoadError::Load(
-                "native runtime symbol not found: skippy_abi_version".to_string(),
-            ));
-        };
-        if !runtime_abi_supported(version) {
-            return Err(NativeRuntimeLoadError::Load(format!(
-                "native runtime ABI {}.{}.{} is not compatible with required ABI {}.{}.{}",
-                version.major,
-                version.minor,
-                version.patch,
-                ABI_VERSION_MAJOR,
-                ABI_VERSION_MINOR,
-                ABI_VERSION_PATCH,
-            )));
-        }
-        Ok(())
-    }
-
-    macro_rules! dynamic_symbols {
-        ($($name:ident($($arg:ident: $arg_ty:ty),* $(,)?) $(-> $ret:ty)?;)+) => {
-            #[allow(non_snake_case)]
-            struct Symbols {
-                _libraries: Vec<Library>,
-                $($name: unsafe extern "C" fn($($arg_ty),*) $(-> $ret)?,)+
-            }
-
-            impl Symbols {
-                unsafe fn load_paths<P>(paths: &[P]) -> Result<Self, NativeRuntimeLoadError>
-                where
-                    P: AsRef<std::path::Path>,
-                {
-                    if paths.is_empty() {
-                        return Err(NativeRuntimeLoadError::Load(
-                            "native runtime did not provide any libraries".to_string(),
-                        ));
-                    }
-                    let mut libraries = Vec::with_capacity(paths.len());
-                    for path in paths {
-                        libraries.push(
-                            unsafe { crate::dynamic_library::load(path.as_ref()) }
-                                .map_err(|err| NativeRuntimeLoadError::Load(err.to_string()))?,
-                        );
-                    }
-                    check_runtime_abi(&libraries)?;
-                    $(
-                        let mut $name = None;
-                        for library in libraries.iter().rev() {
-                            if let Ok(symbol) = unsafe {
-                                library.get::<unsafe extern "C" fn($($arg_ty),*) $(-> $ret)?>(
-                                    concat!(stringify!($name), "\0").as_bytes(),
-                                )
-                            } {
-                                $name = Some(*symbol);
-                                break;
-                            }
-                        };
-                        let $name = $name.ok_or_else(|| {
-                            NativeRuntimeLoadError::Load(format!(
-                                "native runtime symbol not found: {}",
-                                stringify!($name),
-                            ))
-                        })?;
-                    )+
-                    Ok(Self {
-                        _libraries: libraries,
-                        $($name,)+
-                    })
-                }
-            }
-
-            $(
-                #[allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
-                pub unsafe fn $name($($arg: $arg_ty),*) $(-> $ret)? {
-                    unsafe { (symbols().$name)($($arg),*) }
-                }
-            )+
-        };
-    }
-
-    dynamic_symbols! {
-        llama_log_set(log_callback: LlamaLogCallback, user_data: *mut c_void);
-        ggml_log_set(log_callback: LlamaLogCallback, user_data: *mut c_void);
-        llama_model_quantize_default_params() -> LlamaModelQuantizeParams;
-        llama_model_quantize(fname_inp: *const c_char, fname_out: *const c_char, params: *const LlamaModelQuantizeParams) -> u32;
-        skippy_error_free(error: *mut Error);
-        skippy_ngram_cache_create(ngram_min: u16, ngram_max: u16, out_cache: *mut *mut NgramCache, out_error: *mut *mut Error) -> Status;
-        skippy_ngram_cache_free(cache: *mut NgramCache);
-        skippy_ngram_cache_reset(cache: *mut NgramCache, token_ids: *const i32, token_count: usize, out_error: *mut *mut Error) -> Status;
-        skippy_ngram_cache_append(cache: *mut NgramCache, token_ids: *const i32, token_count: usize, out_error: *mut *mut Error) -> Status;
-        skippy_ngram_cache_draft(cache: *mut NgramCache, continuation_prefix: *const i32, continuation_prefix_count: usize, max_draft_tokens: u16, output_tokens: *mut i32, output_token_capacity: usize, out_token_count: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_backend_device_count(out_count: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_backend_device_at(index: usize, out_device: *mut BackendDevice, out_error: *mut *mut Error) -> Status;
-        skippy_model_open(path: *const c_char, config: *const RuntimeConfig, out_model: *mut *mut Model, out_error: *mut *mut Error) -> Status;
-        skippy_model_open_from_parts(paths: *const *const c_char, path_count: usize, config: *const RuntimeConfig, out_model: *mut *mut Model, out_error: *mut *mut Error) -> Status;
-        skippy_model_free(model: *mut Model, out_error: *mut *mut Error) -> Status;
-        skippy_model_llama_model(model: *const Model) -> *const Opaque;
-        skippy_session_create(model: *mut Model, out_session: *mut *mut Session, out_error: *mut *mut Error) -> Status;
-        skippy_session_create_from_resident_prefix(model: *mut Model, cache_seq_id: i32, token_ids: *const i32, token_count: usize, out_session: *mut *mut Session, out_error: *mut *mut Error) -> Status;
-        skippy_session_llama_context(session: *mut Session) -> *mut Opaque;
-        skippy_session_position(session: *const Session) -> i32;
-        skippy_session_batch_size(session: *const Session) -> i32;
-        skippy_session_begin_external_decode(session: *mut Session, out_error: *mut *mut Error) -> Status;
-        skippy_session_end_external_decode(session: *mut Session, out_error: *mut *mut Error) -> Status;
-        skippy_session_set_position(session: *mut Session, n_past: i32, out_error: *mut *mut Error) -> Status;
-        skippy_session_sample_current(session: *mut Session, sampling: *const SamplingConfig, out_predicted_token: *mut i32, out_error: *mut *mut Error) -> Status;
-        skippy_session_configure_chat_sampling(session: *mut Session, sampling: *const SamplingConfig, metadata_json: *const c_char, prompt_token_count: u64, out_error: *mut *mut Error) -> Status;
-        skippy_session_reset(session: *mut Session, out_error: *mut *mut Error) -> Status;
-        skippy_session_free(session: *mut Session, out_error: *mut *mut Error) -> Status;
-        skippy_prefill_chunk(session: *mut Session, token_ids: *const i32, token_count: usize, input_activations: *const c_void, input_activation_bytes: usize, output_activations: *mut c_void, output_activation_capacity: usize, out_output_activation_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_verify_tokens(session: *mut Session, token_ids: *const i32, token_count: usize, output_tokens: *mut i32, output_token_capacity: usize, out_token_count: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_decode_step_sampled(session: *mut Session, token_id: i32, sampling: *const SamplingConfig, input_activation: *const c_void, input_activation_bytes: usize, output_activation: *mut c_void, output_activation_capacity: usize, out_output_activation_bytes: *mut usize, out_predicted_token: *mut i32, out_error: *mut *mut Error) -> Status;
-        skippy_decode_batch_sampled(sessions: *const *mut Session, token_ids: *const i32, sampling: *const *const SamplingConfig, request_count: usize, out_predicted_tokens: *mut i32, predicted_token_capacity: usize, out_error: *mut *mut Error) -> Status;
-        skippy_prefill_chunk_frame(session: *mut Session, token_ids: *const i32, token_count: usize, input_desc: *const ActivationDesc, input_payload: *const c_void, output_desc: *mut ActivationDesc, output_payload: *mut c_void, output_payload_capacity: usize, out_output_payload_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_prefill_chunk_frame_sampled(session: *mut Session, token_ids: *const i32, token_count: usize, sampling: *const SamplingConfig, input_desc: *const ActivationDesc, input_payload: *const c_void, output_desc: *mut ActivationDesc, output_payload: *mut c_void, output_payload_capacity: usize, out_output_payload_bytes: *mut usize, out_predicted_token: *mut i32, out_error: *mut *mut Error) -> Status;
-        skippy_prefill_chunk_frame_with_positions(session: *mut Session, token_ids: *const i32, token_count: usize, positions: *const i32, position_count: usize, input_desc: *const ActivationDesc, input_payload: *const c_void, output_desc: *mut ActivationDesc, output_payload: *mut c_void, output_payload_capacity: usize, out_output_payload_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_prefill_chunk_frame_sampled_with_positions(session: *mut Session, token_ids: *const i32, token_count: usize, positions: *const i32, position_count: usize, sampling: *const SamplingConfig, input_desc: *const ActivationDesc, input_payload: *const c_void, output_desc: *mut ActivationDesc, output_payload: *mut c_void, output_payload_capacity: usize, out_output_payload_bytes: *mut usize, out_predicted_token: *mut i32, out_error: *mut *mut Error) -> Status;
-        skippy_decode_step_frame_sampled(session: *mut Session, token_id: i32, sampling: *const SamplingConfig, input_desc: *const ActivationDesc, input_payload: *const c_void, output_desc: *mut ActivationDesc, output_payload: *mut c_void, output_payload_capacity: usize, out_output_payload_bytes: *mut usize, out_predicted_token: *mut i32, out_error: *mut *mut Error) -> Status;
-        skippy_decode_step_frame_sampled_mtp(session: *mut Session, token_id: i32, sampling: *const SamplingConfig, input_desc: *const ActivationDesc, input_payload: *const c_void, output_desc: *mut ActivationDesc, output_payload: *mut c_void, output_payload_capacity: usize, out_output_payload_bytes: *mut usize, out_predicted_token: *mut i32, max_draft_tokens: usize, out_mtp_draft: *mut NativeMtpDraft, out_error: *mut *mut Error) -> Status;
-        skippy_decode_step_frame_batch_sampled(sessions: *const *mut Session, token_ids: *const i32, sampling: *const *const SamplingConfig, input_descs: *const *const ActivationDesc, input_payloads: *const *const c_void, output_descs: *mut ActivationDesc, output_payloads: *const *mut c_void, output_payload_capacities: *const usize, out_output_payload_bytes: *mut usize, out_predicted_tokens: *mut i32, predicted_token_capacity: usize, request_count: usize, out_error: *mut *mut Error) -> Status;
-        skippy_verify_tokens_frame_sampled(session: *mut Session, token_ids: *const i32, token_count: usize, sampling: *const SamplingConfig, input_desc: *const ActivationDesc, input_payload: *const c_void, output_desc: *mut ActivationDesc, output_payload: *mut c_void, output_payload_capacity: usize, out_output_payload_bytes: *mut usize, output_tokens: *mut i32, output_token_capacity: usize, out_token_count: *mut usize, max_draft_tokens: usize, out_mtp_draft: *mut NativeMtpDraft, out_error: *mut *mut Error) -> Status;
-        skippy_session_copy_output_activation_frame(session: *mut Session, token_count: usize, output_desc: *mut ActivationDesc, output_payload: *mut c_void, output_payload_capacity: usize, out_output_payload_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_session_last_token_signal(session: *mut Session, out_signal: *mut TokenSignal, out_error: *mut *mut Error) -> Status;
-        skippy_session_signal_window(session: *mut Session, window_tokens: u32, out_window: *mut GenerationSignalWindow, out_error: *mut *mut Error) -> Status;
-        skippy_trim_session(session: *mut Session, token_count: u64, out_error: *mut *mut Error) -> Status;
-        skippy_retire_verify_checkpoint(session: *mut Session, token_start: u64, token_count: u64, out_error: *mut *mut Error) -> Status;
-        skippy_export_state(session: *mut Session, layer_start: i32, layer_end: i32, output: *mut c_void, output_capacity: usize, out_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_import_state(session: *mut Session, layer_start: i32, layer_end: i32, input: *const c_void, input_bytes: usize, out_error: *mut *mut Error) -> Status;
-        skippy_export_full_state(session: *mut Session, layer_start: i32, layer_end: i32, output: *mut c_void, output_capacity: usize, out_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_import_full_state(session: *mut Session, layer_start: i32, layer_end: i32, input: *const c_void, input_bytes: usize, out_error: *mut *mut Error) -> Status;
-        skippy_export_kv_page(session: *mut Session, layer_start: i32, layer_end: i32, token_start: u64, token_count: u64, out_desc: *mut KvPageDesc, output: *mut c_void, output_capacity: usize, out_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_import_kv_page(session: *mut Session, desc: *const KvPageDesc, input: *const c_void, input_bytes: usize, out_error: *mut *mut Error) -> Status;
-        skippy_export_recurrent_state(session: *mut Session, output: *mut c_void, output_capacity: usize, out_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_import_recurrent_state(session: *mut Session, input: *const c_void, input_bytes: usize, out_error: *mut *mut Error) -> Status;
-        skippy_session_save_prefix(session: *mut Session, cache_seq_id: i32, token_count: u64, out_error: *mut *mut Error) -> Status;
-        skippy_session_restore_prefix(session: *mut Session, cache_seq_id: i32, token_ids: *const i32, token_count: usize, out_error: *mut *mut Error) -> Status;
-        skippy_session_drop_sequence(session: *mut Session, seq_id: i32, out_error: *mut *mut Error) -> Status;
-        skippy_tokenize(model: *mut Model, text: *const c_char, add_special: bool, output_tokens: *mut i32, output_token_capacity: usize, out_token_count: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_detokenize(model: *mut Model, tokens: *const i32, token_count: usize, output_text: *mut c_char, output_text_capacity: usize, out_text_bytes: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_token_is_eog(model: *mut Model, token_id: i32, out_is_eog: *mut bool, out_error: *mut *mut Error) -> Status;
-        skippy_model_info_open(path: *const c_char, out_info: *mut *mut ModelInfo, out_error: *mut *mut Error) -> Status;
-        skippy_model_info_free(info: *mut ModelInfo, out_error: *mut *mut Error) -> Status;
-        skippy_model_info_tensor_count(info: *mut ModelInfo, out_count: *mut usize, out_error: *mut *mut Error) -> Status;
-        skippy_model_info_tensor_at(info: *mut ModelInfo, index: usize, out_tensor: *mut TensorInfo, out_error: *mut *mut Error) -> Status;
-        skippy_slice_plan_create(info: *mut ModelInfo, out_plan: *mut *mut SlicePlan, out_error: *mut *mut Error) -> Status;
-        skippy_slice_plan_free(plan: *mut SlicePlan, out_error: *mut *mut Error) -> Status;
-        skippy_slice_plan_add_layer_range(plan: *mut SlicePlan, stage_index: i32, layer_start: i32, layer_end: i32, include_embeddings: bool, include_output: bool, out_error: *mut *mut Error) -> Status;
-        skippy_write_slice_gguf(info: *mut ModelInfo, plan: *const SlicePlan, stage_index: i32, output_path: *const c_char, out_error: *mut *mut Error) -> Status;
-        skippy_write_gguf_from_parts(input_paths: *const *const c_char, input_count: usize, output_path: *const c_char, out_error: *mut *mut Error) -> Status;
-        mtmd_default_marker() -> *const c_char;
-        mtmd_helper_log_set(log_callback: LlamaLogCallback, user_data: *mut c_void);
-        mtmd_context_params_default() -> MtmdContextParams;
-        mtmd_init_from_file(mmproj_fname: *const c_char, text_model: *const Opaque, ctx_params: MtmdContextParams) -> *mut MtmdContext;
-        mtmd_free(ctx: *mut MtmdContext);
-        mtmd_helper_bitmap_init_from_buf(ctx: *mut MtmdContext, buf: *const u8, len: usize) -> *mut MtmdBitmap;
-        mtmd_bitmap_free(bitmap: *mut MtmdBitmap);
-        mtmd_input_chunks_init() -> *mut MtmdInputChunks;
-        mtmd_input_chunks_free(chunks: *mut MtmdInputChunks);
-        mtmd_tokenize(ctx: *mut MtmdContext, output: *mut MtmdInputChunks, text: *const MtmdInputText, bitmaps: *const *const MtmdBitmap, n_bitmaps: usize) -> c_int;
-        mtmd_helper_get_n_tokens(chunks: *const MtmdInputChunks) -> usize;
-        mtmd_helper_get_n_pos(chunks: *const MtmdInputChunks) -> i32;
-        mtmd_input_chunks_size(chunks: *const MtmdInputChunks) -> usize;
-        mtmd_input_chunks_get(chunks: *const MtmdInputChunks, index: usize) -> *const Opaque;
-        mtmd_decode_use_mrope(ctx: *const MtmdContext) -> bool;
-        mtmd_input_chunk_get_type(chunk: *const Opaque) -> MtmdInputChunkType;
-        mtmd_input_chunk_get_n_tokens(chunk: *const Opaque) -> usize;
-        mtmd_input_chunk_get_tokens_text(chunk: *const Opaque, out_count: *mut usize) -> *const i32;
-        mtmd_input_chunk_get_tokens_image(chunk: *const Opaque) -> *const Opaque;
-        mtmd_helper_image_get_decoder_pos(image: *const Opaque, pos_0: i32, out_pos: *mut MtmdDecoderPos);
-        mtmd_helper_eval_chunks(ctx: *mut MtmdContext, lctx: *mut Opaque, chunks: *const MtmdInputChunks, n_past: i32, seq_id: i32, n_batch: i32, logits_last: bool, new_n_past: *mut i32) -> c_int;
-        mtmd_helper_eval_chunk_single(ctx: *mut MtmdContext, lctx: *mut Opaque, chunk: *const Opaque, n_past: i32, seq_id: i32, n_batch: i32, logits_last: bool, new_n_past: *mut i32) -> c_int;
-    }
-
-    // -----------------------------------------------------------------------
-    // Optional symbols — not required for library load.
-    // Older runtimes may lack these and callers must check availability first.
-    // -----------------------------------------------------------------------
-
-    type SkippyAbiFeaturesFn = unsafe extern "C" fn() -> u64;
-    type SkippyModelOpenWithEventsFn = unsafe extern "C" fn(
-        path: *const c_char,
-        config: *const RuntimeConfig,
-        reporter: *const SkippyRuntimeEventReporterV1,
-        out_model: *mut *mut Model,
-        out_error: *mut *mut Error,
-    ) -> Status;
-    type SkippyModelOpenFromPartsWithEventsFn = unsafe extern "C" fn(
-        paths: *const *const c_char,
-        path_count: usize,
-        config: *const RuntimeConfig,
-        reporter: *const SkippyRuntimeEventReporterV1,
-        out_model: *mut *mut Model,
-        out_error: *mut *mut Error,
-    ) -> Status;
-    type SkippyApplyChatTemplateJsonFn = unsafe extern "C" fn(
-        model: *mut Model,
-        messages_json: *const c_char,
-        tools_json: *const c_char,
-        tool_choice_json: *const c_char,
-        add_assistant: bool,
-        override_enable_thinking: bool,
-        enable_thinking: bool,
-        parallel_tool_calls: bool,
-        reasoning_format: *const c_char,
-        chat_template_kwargs: *const c_char,
-        output_text: *mut c_char,
-        output_text_capacity: usize,
-        out_text_bytes: *mut usize,
-        output_metadata_json: *mut c_char,
-        output_metadata_json_capacity: usize,
-        out_metadata_json_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-    type SkippyParseChatResponseJsonFn = unsafe extern "C" fn(
-        generated_text: *const c_char,
-        metadata_json: *const c_char,
-        is_partial: bool,
-        output_message_json: *mut c_char,
-        output_message_json_capacity: usize,
-        out_message_json_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    impl Symbols {
-        fn lookup_optional<Sym>(&self, name: &[u8]) -> Option<Sym>
-        where
-            Sym: Copy + 'static,
-        {
-            for library in self._libraries.iter().rev() {
-                if let Ok(sym) = unsafe { library.get::<Sym>(name) } {
-                    return Some(*sym);
-                }
-            }
-            None
-        }
-    }
-
-    pub fn skippy_abi_features_optional() -> Option<SkippyAbiFeaturesFn> {
-        static CACHE: OnceLock<Option<SkippyAbiFeaturesFn>> = OnceLock::new();
-        *CACHE.get_or_init(|| {
-            symbols().lookup_optional::<SkippyAbiFeaturesFn>(b"skippy_abi_features\0")
-        })
-    }
-
-    pub fn skippy_model_open_with_events_fn() -> Option<SkippyModelOpenWithEventsFn> {
-        static CACHE: OnceLock<Option<SkippyModelOpenWithEventsFn>> = OnceLock::new();
-        *CACHE.get_or_init(|| {
-            symbols()
-                .lookup_optional::<SkippyModelOpenWithEventsFn>(b"skippy_model_open_with_events\0")
-        })
-    }
-
-    pub fn skippy_model_attach_mtp_draft_model_fn() -> Option<SkippyModelAttachMtpDraftModelFn> {
-        static CACHE: OnceLock<Option<SkippyModelAttachMtpDraftModelFn>> = OnceLock::new();
-        *CACHE.get_or_init(|| {
-            symbols().lookup_optional::<SkippyModelAttachMtpDraftModelFn>(
-                b"skippy_model_attach_mtp_draft_model\0",
-            )
-        })
-    }
-
-    pub fn skippy_decode_step_sampled_mtp_fn() -> Option<SkippyDecodeStepSampledMtpFn> {
-        static CACHE: OnceLock<Option<SkippyDecodeStepSampledMtpFn>> = OnceLock::new();
-        *CACHE.get_or_init(|| {
-            symbols().lookup_optional::<SkippyDecodeStepSampledMtpFn>(
-                b"skippy_decode_step_sampled_mtp\0",
-            )
-        })
-    }
-
-    pub fn skippy_model_open_from_parts_with_events_fn()
-    -> Option<SkippyModelOpenFromPartsWithEventsFn> {
-        static CACHE: OnceLock<Option<SkippyModelOpenFromPartsWithEventsFn>> = OnceLock::new();
-        *CACHE.get_or_init(|| {
-            symbols().lookup_optional::<SkippyModelOpenFromPartsWithEventsFn>(
-                b"skippy_model_open_from_parts_with_events\0",
-            )
-        })
-    }
-
-    fn skippy_apply_chat_template_json_fn() -> Option<SkippyApplyChatTemplateJsonFn> {
-        static CACHE: OnceLock<Option<SkippyApplyChatTemplateJsonFn>> = OnceLock::new();
-        *CACHE.get_or_init(|| {
-            symbols().lookup_optional::<SkippyApplyChatTemplateJsonFn>(
-                b"skippy_apply_chat_template_json\0",
-            )
-        })
-    }
-
-    fn skippy_parse_chat_response_json_fn() -> Option<SkippyParseChatResponseJsonFn> {
-        static CACHE: OnceLock<Option<SkippyParseChatResponseJsonFn>> = OnceLock::new();
-        *CACHE.get_or_init(|| {
-            symbols().lookup_optional::<SkippyParseChatResponseJsonFn>(
-                b"skippy_parse_chat_response_json\0",
-            )
-        })
-    }
-
-    #[allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
-    pub unsafe fn skippy_apply_chat_template_json(
-        model: *mut Model,
-        messages_json: *const c_char,
-        tools_json: *const c_char,
-        tool_choice_json: *const c_char,
-        add_assistant: bool,
-        override_enable_thinking: bool,
-        enable_thinking: bool,
-        parallel_tool_calls: bool,
-        reasoning_format: *const c_char,
-        chat_template_kwargs: *const c_char,
-        output_text: *mut c_char,
-        output_text_capacity: usize,
-        out_text_bytes: *mut usize,
-        output_metadata_json: *mut c_char,
-        output_metadata_json_capacity: usize,
-        out_metadata_json_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status {
-        let Some(function) = skippy_apply_chat_template_json_fn() else {
-            return Status::Unsupported;
-        };
-        unsafe {
-            function(
-                model,
-                messages_json,
-                tools_json,
-                tool_choice_json,
-                add_assistant,
-                override_enable_thinking,
-                enable_thinking,
-                parallel_tool_calls,
-                reasoning_format,
-                chat_template_kwargs,
-                output_text,
-                output_text_capacity,
-                out_text_bytes,
-                output_metadata_json,
-                output_metadata_json_capacity,
-                out_metadata_json_bytes,
-                out_error,
-            )
-        }
-    }
-
-    #[allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
-    pub unsafe fn skippy_parse_chat_response_json(
-        generated_text: *const c_char,
-        metadata_json: *const c_char,
-        is_partial: bool,
-        output_message_json: *mut c_char,
-        output_message_json_capacity: usize,
-        out_message_json_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status {
-        let Some(function) = skippy_parse_chat_response_json_fn() else {
-            return Status::Unsupported;
-        };
-        unsafe {
-            function(
-                generated_text,
-                metadata_json,
-                is_partial,
-                output_message_json,
-                output_message_json_capacity,
-                out_message_json_bytes,
-                out_error,
-            )
-        }
-    }
-
-    #[allow(clippy::missing_safety_doc, clippy::too_many_arguments)]
-    pub unsafe fn skippy_decode_step_sampled_mtp(
-        session: *mut Session,
-        token_id: i32,
-        sampling: *const SamplingConfig,
-        out_predicted_token: *mut i32,
-        max_draft_tokens: usize,
-        out_mtp_draft: *mut NativeMtpDraft,
-        out_error: *mut *mut Error,
-    ) -> Status {
-        let Some(function) = skippy_decode_step_sampled_mtp_fn() else {
-            return Status::Unsupported;
-        };
-        unsafe {
-            function(
-                session,
-                token_id,
-                sampling,
-                out_predicted_token,
-                max_draft_tokens,
-                out_mtp_draft,
-                out_error,
-            )
-        }
-    }
-}
-
-#[cfg(feature = "dynamic-runtime")]
-pub use dynamic::*;
-
-#[cfg(feature = "dynamic-runtime")]
-/// Returns the skippy ABI feature bitmask.
-/// Requires the native runtime to be loaded first (checked by caller).
-pub fn skippy_abi_features() -> u64 {
-    try_abi_features().expect("skippy_abi_features not available in loaded runtime")
-}
-
-/// Returns the Skippy ABI feature bitmask when the loaded dynamic runtime
-/// exports feature probing.
-#[cfg(feature = "dynamic-runtime")]
-pub fn try_abi_features() -> Option<u64> {
-    dynamic::skippy_abi_features_optional().map(|features| unsafe { features() })
-}
-
-/// Returns the active Skippy ABI feature bitmask through a safe Rust wrapper.
-#[cfg(feature = "dynamic-runtime")]
-pub fn abi_features() -> u64 {
-    skippy_abi_features()
-}
-
-/// Returns the statically linked Skippy ABI feature bitmask.
+mod dynamic;
+mod model;
+mod multimodal;
+mod runtime;
+mod sampling;
+mod state;
 #[cfg(not(feature = "dynamic-runtime"))]
-pub fn try_abi_features() -> Option<u64> {
-    // SAFETY: the statically linked ABI exposes this nullary query with no
-    // caller-owned pointers or lifetime requirements.
-    Some(unsafe { skippy_abi_features() })
-}
-
-/// Returns the statically linked Skippy ABI feature bitmask.
-#[cfg(not(feature = "dynamic-runtime"))]
-pub fn abi_features() -> u64 {
-    // SAFETY: the statically linked ABI exposes this nullary query with no
-    // caller-owned pointers or lifetime requirements.
-    unsafe { skippy_abi_features() }
-}
-
-#[cfg(not(feature = "dynamic-runtime"))]
-unsafe extern "C" {
-    pub fn llama_log_set(log_callback: LlamaLogCallback, user_data: *mut c_void);
-
-    pub fn ggml_log_set(log_callback: LlamaLogCallback, user_data: *mut c_void);
-
-    pub fn llama_model_quantize_default_params() -> LlamaModelQuantizeParams;
-
-    pub fn llama_model_quantize(
-        fname_inp: *const c_char,
-        fname_out: *const c_char,
-        params: *const LlamaModelQuantizeParams,
-    ) -> u32;
-
-    pub fn skippy_abi_features() -> u64;
-
-    pub fn skippy_error_free(error: *mut Error);
-
-    pub fn skippy_ngram_cache_create(
-        ngram_min: u16,
-        ngram_max: u16,
-        out_cache: *mut *mut NgramCache,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_ngram_cache_free(cache: *mut NgramCache);
-
-    pub fn skippy_ngram_cache_reset(
-        cache: *mut NgramCache,
-        token_ids: *const i32,
-        token_count: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_ngram_cache_append(
-        cache: *mut NgramCache,
-        token_ids: *const i32,
-        token_count: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_ngram_cache_draft(
-        cache: *mut NgramCache,
-        continuation_prefix: *const i32,
-        continuation_prefix_count: usize,
-        max_draft_tokens: u16,
-        output_tokens: *mut i32,
-        output_token_capacity: usize,
-        out_token_count: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_backend_device_count(out_count: *mut usize, out_error: *mut *mut Error)
-    -> Status;
-
-    pub fn skippy_backend_device_at(
-        index: usize,
-        out_device: *mut BackendDevice,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_model_open(
-        path: *const c_char,
-        config: *const RuntimeConfig,
-        out_model: *mut *mut Model,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_model_open_from_parts(
-        paths: *const *const c_char,
-        path_count: usize,
-        config: *const RuntimeConfig,
-        out_model: *mut *mut Model,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_model_attach_mtp_draft_model(
-        target_model: *mut Model,
-        path: *const c_char,
-        config: *const RuntimeConfig,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_model_free(model: *mut Model, out_error: *mut *mut Error) -> Status;
-
-    pub fn skippy_model_llama_model(model: *const Model) -> *const Opaque;
-
-    pub fn skippy_session_create(
-        model: *mut Model,
-        out_session: *mut *mut Session,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_create_from_resident_prefix(
-        model: *mut Model,
-        cache_seq_id: i32,
-        token_ids: *const i32,
-        token_count: usize,
-        out_session: *mut *mut Session,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_llama_context(session: *mut Session) -> *mut Opaque;
-
-    pub fn skippy_session_position(session: *const Session) -> i32;
-
-    pub fn skippy_session_batch_size(session: *const Session) -> i32;
-
-    pub fn skippy_session_begin_external_decode(
-        session: *mut Session,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_end_external_decode(
-        session: *mut Session,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_set_position(
-        session: *mut Session,
-        n_past: i32,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_sample_current(
-        session: *mut Session,
-        sampling: *const SamplingConfig,
-        out_predicted_token: *mut i32,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_configure_chat_sampling(
-        session: *mut Session,
-        sampling: *const SamplingConfig,
-        metadata_json: *const c_char,
-        prompt_token_count: u64,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_reset(session: *mut Session, out_error: *mut *mut Error) -> Status;
-
-    pub fn skippy_session_free(session: *mut Session, out_error: *mut *mut Error) -> Status;
-
-    pub fn skippy_prefill_chunk(
-        session: *mut Session,
-        token_ids: *const i32,
-        token_count: usize,
-        input_activations: *const c_void,
-        input_activation_bytes: usize,
-        output_activations: *mut c_void,
-        output_activation_capacity: usize,
-        out_output_activation_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_verify_tokens(
-        session: *mut Session,
-        token_ids: *const i32,
-        token_count: usize,
-        output_tokens: *mut i32,
-        output_token_capacity: usize,
-        out_token_count: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_decode_step_sampled(
-        session: *mut Session,
-        token_id: i32,
-        sampling: *const SamplingConfig,
-        input_activation: *const c_void,
-        input_activation_bytes: usize,
-        output_activation: *mut c_void,
-        output_activation_capacity: usize,
-        out_output_activation_bytes: *mut usize,
-        out_predicted_token: *mut i32,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_decode_step_sampled_mtp(
-        session: *mut Session,
-        token_id: i32,
-        sampling: *const SamplingConfig,
-        out_predicted_token: *mut i32,
-        max_draft_tokens: usize,
-        out_mtp_draft: *mut NativeMtpDraft,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_decode_batch_sampled(
-        sessions: *const *mut Session,
-        token_ids: *const i32,
-        sampling: *const *const SamplingConfig,
-        request_count: usize,
-        out_predicted_tokens: *mut i32,
-        predicted_token_capacity: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_prefill_chunk_frame(
-        session: *mut Session,
-        token_ids: *const i32,
-        token_count: usize,
-        input_desc: *const ActivationDesc,
-        input_payload: *const c_void,
-        output_desc: *mut ActivationDesc,
-        output_payload: *mut c_void,
-        output_payload_capacity: usize,
-        out_output_payload_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_prefill_chunk_frame_sampled(
-        session: *mut Session,
-        token_ids: *const i32,
-        token_count: usize,
-        sampling: *const SamplingConfig,
-        input_desc: *const ActivationDesc,
-        input_payload: *const c_void,
-        output_desc: *mut ActivationDesc,
-        output_payload: *mut c_void,
-        output_payload_capacity: usize,
-        out_output_payload_bytes: *mut usize,
-        out_predicted_token: *mut i32,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_prefill_chunk_frame_with_positions(
-        session: *mut Session,
-        token_ids: *const i32,
-        token_count: usize,
-        positions: *const i32,
-        position_count: usize,
-        input_desc: *const ActivationDesc,
-        input_payload: *const c_void,
-        output_desc: *mut ActivationDesc,
-        output_payload: *mut c_void,
-        output_payload_capacity: usize,
-        out_output_payload_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_prefill_chunk_frame_sampled_with_positions(
-        session: *mut Session,
-        token_ids: *const i32,
-        token_count: usize,
-        positions: *const i32,
-        position_count: usize,
-        sampling: *const SamplingConfig,
-        input_desc: *const ActivationDesc,
-        input_payload: *const c_void,
-        output_desc: *mut ActivationDesc,
-        output_payload: *mut c_void,
-        output_payload_capacity: usize,
-        out_output_payload_bytes: *mut usize,
-        out_predicted_token: *mut i32,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_verify_tokens_frame_sampled(
-        session: *mut Session,
-        token_ids: *const i32,
-        token_count: usize,
-        sampling: *const SamplingConfig,
-        input_desc: *const ActivationDesc,
-        input_payload: *const c_void,
-        output_desc: *mut ActivationDesc,
-        output_payload: *mut c_void,
-        output_payload_capacity: usize,
-        out_output_payload_bytes: *mut usize,
-        output_tokens: *mut i32,
-        output_token_capacity: usize,
-        out_token_count: *mut usize,
-        max_draft_tokens: usize,
-        out_mtp_draft: *mut NativeMtpDraft,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_decode_step_frame_sampled(
-        session: *mut Session,
-        token_id: i32,
-        sampling: *const SamplingConfig,
-        input_desc: *const ActivationDesc,
-        input_payload: *const c_void,
-        output_desc: *mut ActivationDesc,
-        output_payload: *mut c_void,
-        output_payload_capacity: usize,
-        out_output_payload_bytes: *mut usize,
-        out_predicted_token: *mut i32,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_decode_step_frame_sampled_mtp(
-        session: *mut Session,
-        token_id: i32,
-        sampling: *const SamplingConfig,
-        input_desc: *const ActivationDesc,
-        input_payload: *const c_void,
-        output_desc: *mut ActivationDesc,
-        output_payload: *mut c_void,
-        output_payload_capacity: usize,
-        out_output_payload_bytes: *mut usize,
-        out_predicted_token: *mut i32,
-        max_draft_tokens: usize,
-        out_mtp_draft: *mut NativeMtpDraft,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_decode_step_frame_batch_sampled(
-        sessions: *const *mut Session,
-        token_ids: *const i32,
-        sampling: *const *const SamplingConfig,
-        input_descs: *const *const ActivationDesc,
-        input_payloads: *const *const c_void,
-        output_descs: *mut ActivationDesc,
-        output_payloads: *const *mut c_void,
-        output_payload_capacities: *const usize,
-        out_output_payload_bytes: *mut usize,
-        out_predicted_tokens: *mut i32,
-        predicted_token_capacity: usize,
-        request_count: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_copy_output_activation_frame(
-        session: *mut Session,
-        token_count: usize,
-        output_desc: *mut ActivationDesc,
-        output_payload: *mut c_void,
-        output_payload_capacity: usize,
-        out_output_payload_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_last_token_signal(
-        session: *mut Session,
-        out_signal: *mut TokenSignal,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_signal_window(
-        session: *mut Session,
-        window_tokens: u32,
-        out_window: *mut GenerationSignalWindow,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_trim_session(
-        session: *mut Session,
-        token_count: u64,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_retire_verify_checkpoint(
-        session: *mut Session,
-        token_start: u64,
-        token_count: u64,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_export_state(
-        session: *mut Session,
-        layer_start: i32,
-        layer_end: i32,
-        output: *mut c_void,
-        output_capacity: usize,
-        out_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_import_state(
-        session: *mut Session,
-        layer_start: i32,
-        layer_end: i32,
-        input: *const c_void,
-        input_bytes: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_export_full_state(
-        session: *mut Session,
-        layer_start: i32,
-        layer_end: i32,
-        output: *mut c_void,
-        output_capacity: usize,
-        out_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_import_full_state(
-        session: *mut Session,
-        layer_start: i32,
-        layer_end: i32,
-        input: *const c_void,
-        input_bytes: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_export_kv_page(
-        session: *mut Session,
-        layer_start: i32,
-        layer_end: i32,
-        token_start: u64,
-        token_count: u64,
-        out_desc: *mut KvPageDesc,
-        output: *mut c_void,
-        output_capacity: usize,
-        out_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_import_kv_page(
-        session: *mut Session,
-        desc: *const KvPageDesc,
-        input: *const c_void,
-        input_bytes: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_export_recurrent_state(
-        session: *mut Session,
-        output: *mut c_void,
-        output_capacity: usize,
-        out_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_import_recurrent_state(
-        session: *mut Session,
-        input: *const c_void,
-        input_bytes: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_save_prefix(
-        session: *mut Session,
-        cache_seq_id: i32,
-        token_count: u64,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_restore_prefix(
-        session: *mut Session,
-        cache_seq_id: i32,
-        token_ids: *const i32,
-        token_count: usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_session_drop_sequence(
-        session: *mut Session,
-        seq_id: i32,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_tokenize(
-        model: *mut Model,
-        text: *const c_char,
-        add_special: bool,
-        output_tokens: *mut i32,
-        output_token_capacity: usize,
-        out_token_count: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_detokenize(
-        model: *mut Model,
-        tokens: *const i32,
-        token_count: usize,
-        output_text: *mut c_char,
-        output_text_capacity: usize,
-        out_text_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_token_is_eog(
-        model: *mut Model,
-        token_id: i32,
-        out_is_eog: *mut bool,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_apply_chat_template_json(
-        model: *mut Model,
-        messages_json: *const c_char,
-        tools_json: *const c_char,
-        tool_choice_json: *const c_char,
-        add_assistant: bool,
-        override_enable_thinking: bool,
-        enable_thinking: bool,
-        parallel_tool_calls: bool,
-        reasoning_format: *const c_char,
-        chat_template_kwargs: *const c_char,
-        output_text: *mut c_char,
-        output_text_capacity: usize,
-        out_text_bytes: *mut usize,
-        output_metadata_json: *mut c_char,
-        output_metadata_json_capacity: usize,
-        out_metadata_json_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_parse_chat_response_json(
-        generated_text: *const c_char,
-        metadata_json: *const c_char,
-        is_partial: bool,
-        output_message_json: *mut c_char,
-        output_message_json_capacity: usize,
-        out_message_json_bytes: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_model_info_open(
-        path: *const c_char,
-        out_info: *mut *mut ModelInfo,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_model_info_free(info: *mut ModelInfo, out_error: *mut *mut Error) -> Status;
-
-    pub fn skippy_model_info_tensor_count(
-        info: *mut ModelInfo,
-        out_count: *mut usize,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_model_info_tensor_at(
-        info: *mut ModelInfo,
-        index: usize,
-        out_tensor: *mut TensorInfo,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_slice_plan_create(
-        info: *mut ModelInfo,
-        out_plan: *mut *mut SlicePlan,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_slice_plan_free(plan: *mut SlicePlan, out_error: *mut *mut Error) -> Status;
-
-    pub fn skippy_slice_plan_add_layer_range(
-        plan: *mut SlicePlan,
-        stage_index: i32,
-        layer_start: i32,
-        layer_end: i32,
-        include_embeddings: bool,
-        include_output: bool,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_write_slice_gguf(
-        info: *mut ModelInfo,
-        plan: *const SlicePlan,
-        stage_index: i32,
-        output_path: *const c_char,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn skippy_write_gguf_from_parts(
-        input_paths: *const *const c_char,
-        input_count: usize,
-        output_path: *const c_char,
-        out_error: *mut *mut Error,
-    ) -> Status;
-
-    pub fn mtmd_default_marker() -> *const c_char;
-
-    pub fn mtmd_helper_log_set(log_callback: LlamaLogCallback, user_data: *mut c_void);
-
-    pub fn mtmd_context_params_default() -> MtmdContextParams;
-
-    pub fn mtmd_init_from_file(
-        mmproj_fname: *const c_char,
-        text_model: *const Opaque,
-        ctx_params: MtmdContextParams,
-    ) -> *mut MtmdContext;
-
-    pub fn mtmd_free(ctx: *mut MtmdContext);
-
-    pub fn mtmd_helper_bitmap_init_from_buf(
-        ctx: *mut MtmdContext,
-        buf: *const u8,
-        len: usize,
-    ) -> *mut MtmdBitmap;
-
-    pub fn mtmd_bitmap_free(bitmap: *mut MtmdBitmap);
-
-    pub fn mtmd_input_chunks_init() -> *mut MtmdInputChunks;
-
-    pub fn mtmd_input_chunks_free(chunks: *mut MtmdInputChunks);
-
-    pub fn mtmd_tokenize(
-        ctx: *mut MtmdContext,
-        output: *mut MtmdInputChunks,
-        text: *const MtmdInputText,
-        bitmaps: *const *const MtmdBitmap,
-        n_bitmaps: usize,
-    ) -> c_int;
-
-    pub fn mtmd_helper_get_n_tokens(chunks: *const MtmdInputChunks) -> usize;
-
-    pub fn mtmd_helper_get_n_pos(chunks: *const MtmdInputChunks) -> i32;
-
-    pub fn mtmd_input_chunks_size(chunks: *const MtmdInputChunks) -> usize;
-
-    pub fn mtmd_input_chunks_get(chunks: *const MtmdInputChunks, index: usize) -> *const Opaque;
-
-    pub fn mtmd_decode_use_mrope(ctx: *const MtmdContext) -> bool;
-
-    pub fn mtmd_input_chunk_get_type(chunk: *const Opaque) -> MtmdInputChunkType;
-
-    pub fn mtmd_input_chunk_get_n_tokens(chunk: *const Opaque) -> usize;
-
-    pub fn mtmd_input_chunk_get_tokens_text(
-        chunk: *const Opaque,
-        out_count: *mut usize,
-    ) -> *const i32;
-
-    pub fn mtmd_input_chunk_get_tokens_image(chunk: *const Opaque) -> *const Opaque;
-
-    pub fn mtmd_helper_image_get_decoder_pos(
-        image: *const Opaque,
-        pos_0: i32,
-        out_pos: *mut MtmdDecoderPos,
-    );
-
-    pub fn mtmd_helper_eval_chunks(
-        ctx: *mut MtmdContext,
-        lctx: *mut Opaque,
-        chunks: *const MtmdInputChunks,
-        n_past: i32,
-        seq_id: i32,
-        n_batch: i32,
-        logits_last: bool,
-        new_n_past: *mut i32,
-    ) -> c_int;
-
-    pub fn mtmd_helper_eval_chunk_single(
-        ctx: *mut MtmdContext,
-        lctx: *mut Opaque,
-        chunk: *const Opaque,
-        n_past: i32,
-        seq_id: i32,
-        n_batch: i32,
-        logits_last: bool,
-        new_n_past: *mut i32,
-    ) -> c_int;
-}
-
-#[cfg(not(feature = "dynamic-runtime"))]
-pub fn skippy_model_attach_mtp_draft_model_fn() -> Option<SkippyModelAttachMtpDraftModelFn> {
-    Some(skippy_model_attach_mtp_draft_model)
-}
-
-#[cfg(not(feature = "dynamic-runtime"))]
-pub fn skippy_decode_step_sampled_mtp_fn() -> Option<SkippyDecodeStepSampledMtpFn> {
-    Some(skippy_decode_step_sampled_mtp)
-}
-
-pub type Opaque = c_void;
+mod static_bindings;
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use std::mem::{offset_of, size_of};
+mod tests;
 
-    const fn version(major: u32, minor: u32, patch: u32) -> AbiVersion {
-        AbiVersion {
-            major,
-            minor,
-            patch,
-        }
-    }
+pub use abi::{
+    ABI_VERSION_MAJOR, ABI_VERSION_MINOR, ABI_VERSION_PATCH, AbiVersion, ActivationDType,
+    ActivationLayout, BACKEND_DEVICE_CAP_ASYNC, BACKEND_DEVICE_CAP_BUFFER_FROM_HOST_PTR,
+    BACKEND_DEVICE_CAP_EVENTS, BACKEND_DEVICE_CAP_HOST_BUFFER, BackendDevice, BackendDeviceType,
+    Error, FEATURE_BACKEND_DEVICES, FEATURE_INKLING_MTP_MM, FEATURE_NATIVE_MTP_N1,
+    FEATURE_NGRAM_CACHE_DRAFT, FEATURE_RUNTIME_EVENTS, LlamaLogCallback, LoadMode, Model,
+    ModelInfo, MtmdProgressCallback, MtpSource, NgramCache, Opaque, RuntimeConfig, Session,
+    SkippyDecodeStepSampledMtpFn, SkippyModelAttachMtpDraftModelFn, SkippyRuntimeEventCallback,
+    SkippyRuntimeEventCategory, SkippyRuntimeEventEmitterKind, SkippyRuntimeEventFailureCode,
+    SkippyRuntimeEventKind, SkippyRuntimeEventProgressUnit, SkippyRuntimeEventReporterV1,
+    SkippyRuntimeEventV1, SlicePlan, Status, TensorRole, runtime_abi_supported,
+};
+pub use activation::{ACTIVATION_FLAG_INKLING_MTP_EMBD, ActivationDesc, LogitBias, TensorInfo};
+pub use model::{
+    GgmlType, LlamaFileType, LlamaModelImatrixData, LlamaModelKvOverride, LlamaModelKvOverrideType,
+    LlamaModelKvOverrideValue, LlamaModelQuantizeParams, LlamaModelTensorOverride,
+};
+pub use multimodal::{
+    MtmdBitmap, MtmdContext, MtmdContextParams, MtmdDecoderPos, MtmdInputChunkType,
+    MtmdInputChunks, MtmdInputText,
+};
+pub use runtime::{NativeRuntimeLoadError, abi_features, try_abi_features};
+pub use sampling::{
+    GenerationSignalWindow, NATIVE_MTP_MAX_DRAFT_TOKENS, NativeMtpDraft, SamplingConfig,
+    TokenSignal,
+};
+pub use state::{
+    KV_PAGE_CODEC_ISWA_COMPOSITE_V1, KV_PAGE_CODEC_SINGLE_V1, KvPageComponentDesc, KvPageDesc,
+};
 
-    #[test]
-    fn accepts_current_patch_runtime() {
-        assert!(runtime_abi_supported(version(
-            ABI_VERSION_MAJOR,
-            ABI_VERSION_MINOR,
-            ABI_VERSION_PATCH,
-        )));
-    }
+#[cfg(not(feature = "dynamic-runtime"))]
+pub use runtime::{
+    load_native_runtime_libraries, load_native_runtime_library, native_runtime_loaded,
+    skippy_decode_step_sampled_mtp_fn, skippy_model_attach_mtp_draft_model_fn,
+};
 
-    #[test]
-    fn rejects_other_patch_runtimes() {
-        assert!(!runtime_abi_supported(version(
-            ABI_VERSION_MAJOR,
-            ABI_VERSION_MINOR,
-            ABI_VERSION_PATCH + 1,
-        )));
-        assert!(!runtime_abi_supported(version(
-            ABI_VERSION_MAJOR,
-            ABI_VERSION_MINOR,
-            ABI_VERSION_PATCH - 1,
-        )));
-    }
+#[cfg(feature = "dynamic-runtime")]
+pub use runtime::skippy_abi_features;
 
-    #[test]
-    fn rejects_major_and_minor_mismatches() {
-        assert!(!runtime_abi_supported(version(
-            ABI_VERSION_MAJOR + 1,
-            ABI_VERSION_MINOR,
-            ABI_VERSION_PATCH,
-        )));
-        assert!(!runtime_abi_supported(version(
-            ABI_VERSION_MAJOR,
-            ABI_VERSION_MINOR + 1,
-            ABI_VERSION_PATCH,
-        )));
-    }
+#[cfg(feature = "dynamic-runtime")]
+pub use dynamic::{
+    ggml_log_set, llama_log_set, llama_model_quantize, llama_model_quantize_default_params,
+    load_native_runtime_libraries, load_native_runtime_library, mtmd_bitmap_free,
+    mtmd_context_params_default, mtmd_decode_use_mrope, mtmd_default_marker, mtmd_free,
+    mtmd_helper_bitmap_init_from_buf, mtmd_helper_eval_chunk_single, mtmd_helper_eval_chunks,
+    mtmd_helper_get_n_pos, mtmd_helper_get_n_tokens, mtmd_helper_image_get_decoder_pos,
+    mtmd_helper_log_set, mtmd_init_from_file, mtmd_input_chunk_get_n_tokens,
+    mtmd_input_chunk_get_tokens_image, mtmd_input_chunk_get_tokens_text, mtmd_input_chunk_get_type,
+    mtmd_input_chunks_free, mtmd_input_chunks_get, mtmd_input_chunks_init, mtmd_input_chunks_size,
+    mtmd_tokenize, native_runtime_loaded, skippy_abi_features_optional,
+    skippy_apply_chat_template_json, skippy_backend_device_at, skippy_backend_device_count,
+    skippy_decode_batch_sampled, skippy_decode_step_frame_batch_sampled,
+    skippy_decode_step_frame_sampled, skippy_decode_step_frame_sampled_mtp,
+    skippy_decode_step_sampled, skippy_decode_step_sampled_mtp, skippy_decode_step_sampled_mtp_fn,
+    skippy_detokenize, skippy_error_free, skippy_export_full_state, skippy_export_kv_page,
+    skippy_export_recurrent_state, skippy_export_state, skippy_import_full_state,
+    skippy_import_kv_page, skippy_import_recurrent_state, skippy_import_state,
+    skippy_model_attach_mtp_draft_model_fn, skippy_model_free, skippy_model_info_free,
+    skippy_model_info_open, skippy_model_info_tensor_at, skippy_model_info_tensor_count,
+    skippy_model_llama_model, skippy_model_open, skippy_model_open_from_parts,
+    skippy_model_open_from_parts_with_events_fn, skippy_model_open_with_events_fn,
+    skippy_ngram_cache_append, skippy_ngram_cache_create, skippy_ngram_cache_draft,
+    skippy_ngram_cache_free, skippy_ngram_cache_reset, skippy_parse_chat_response_json,
+    skippy_prefill_chunk, skippy_prefill_chunk_frame, skippy_prefill_chunk_frame_sampled,
+    skippy_prefill_chunk_frame_sampled_with_positions, skippy_prefill_chunk_frame_with_positions,
+    skippy_retire_verify_checkpoint, skippy_session_batch_size,
+    skippy_session_begin_external_decode, skippy_session_configure_chat_sampling,
+    skippy_session_copy_output_activation_frame, skippy_session_create,
+    skippy_session_create_from_resident_prefix, skippy_session_drop_sequence,
+    skippy_session_end_external_decode, skippy_session_free, skippy_session_last_token_signal,
+    skippy_session_llama_context, skippy_session_position, skippy_session_reset,
+    skippy_session_restore_prefix, skippy_session_sample_current, skippy_session_save_prefix,
+    skippy_session_set_position, skippy_session_signal_window, skippy_slice_plan_add_layer_range,
+    skippy_slice_plan_create, skippy_slice_plan_free, skippy_token_is_eog, skippy_tokenize,
+    skippy_trim_session, skippy_verify_tokens, skippy_verify_tokens_frame_sampled,
+    skippy_write_gguf_from_parts, skippy_write_slice_gguf,
+};
 
-    #[test]
-    #[cfg(target_pointer_width = "64")]
-    fn mtmd_context_params_matches_native_layout() {
-        assert_eq!(size_of::<MtmdContextParams>(), 80);
-        assert_eq!(offset_of!(MtmdContextParams, batch_max_tokens), 56);
-        assert_eq!(offset_of!(MtmdContextParams, progress_callback), 64);
-        assert_eq!(
-            offset_of!(MtmdContextParams, progress_callback_user_data),
-            72
-        );
-    }
-
-    #[test]
-    #[cfg(not(feature = "dynamic-runtime"))]
-    fn native_mtmd_defaults_cross_the_ffi_boundary() {
-        let params = unsafe { mtmd_context_params_default() };
-
-        assert_eq!(params.batch_max_tokens, 1024);
-        assert!(params.progress_callback.is_none());
-        assert!(params.progress_callback_user_data.is_null());
-    }
-}
+#[cfg(not(feature = "dynamic-runtime"))]
+pub use static_bindings::{
+    ggml_log_set, llama_log_set, llama_model_quantize, llama_model_quantize_default_params,
+    mtmd_bitmap_free, mtmd_context_params_default, mtmd_decode_use_mrope, mtmd_default_marker,
+    mtmd_free, mtmd_helper_bitmap_init_from_buf, mtmd_helper_eval_chunk_single,
+    mtmd_helper_eval_chunks, mtmd_helper_get_n_pos, mtmd_helper_get_n_tokens,
+    mtmd_helper_image_get_decoder_pos, mtmd_helper_log_set, mtmd_init_from_file,
+    mtmd_input_chunk_get_n_tokens, mtmd_input_chunk_get_tokens_image,
+    mtmd_input_chunk_get_tokens_text, mtmd_input_chunk_get_type, mtmd_input_chunks_free,
+    mtmd_input_chunks_get, mtmd_input_chunks_init, mtmd_input_chunks_size, mtmd_tokenize,
+    skippy_abi_features, skippy_apply_chat_template_json, skippy_backend_device_at,
+    skippy_backend_device_count, skippy_decode_batch_sampled,
+    skippy_decode_step_frame_batch_sampled, skippy_decode_step_frame_sampled,
+    skippy_decode_step_frame_sampled_mtp, skippy_decode_step_sampled,
+    skippy_decode_step_sampled_mtp, skippy_detokenize, skippy_error_free, skippy_export_full_state,
+    skippy_export_kv_page, skippy_export_recurrent_state, skippy_export_state,
+    skippy_import_full_state, skippy_import_kv_page, skippy_import_recurrent_state,
+    skippy_import_state, skippy_model_attach_mtp_draft_model, skippy_model_free,
+    skippy_model_info_free, skippy_model_info_open, skippy_model_info_tensor_at,
+    skippy_model_info_tensor_count, skippy_model_llama_model, skippy_model_open,
+    skippy_model_open_from_parts, skippy_ngram_cache_append, skippy_ngram_cache_create,
+    skippy_ngram_cache_draft, skippy_ngram_cache_free, skippy_ngram_cache_reset,
+    skippy_parse_chat_response_json, skippy_prefill_chunk, skippy_prefill_chunk_frame,
+    skippy_prefill_chunk_frame_sampled, skippy_prefill_chunk_frame_sampled_with_positions,
+    skippy_prefill_chunk_frame_with_positions, skippy_retire_verify_checkpoint,
+    skippy_session_batch_size, skippy_session_begin_external_decode,
+    skippy_session_configure_chat_sampling, skippy_session_copy_output_activation_frame,
+    skippy_session_create, skippy_session_create_from_resident_prefix,
+    skippy_session_drop_sequence, skippy_session_end_external_decode, skippy_session_free,
+    skippy_session_last_token_signal, skippy_session_llama_context, skippy_session_position,
+    skippy_session_reset, skippy_session_restore_prefix, skippy_session_sample_current,
+    skippy_session_save_prefix, skippy_session_set_position, skippy_session_signal_window,
+    skippy_slice_plan_add_layer_range, skippy_slice_plan_create, skippy_slice_plan_free,
+    skippy_token_is_eog, skippy_tokenize, skippy_trim_session, skippy_verify_tokens,
+    skippy_verify_tokens_frame_sampled, skippy_write_gguf_from_parts, skippy_write_slice_gguf,
+};
